pyproject.toml
[project]
name = "adk-site-builder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.24.1",
    "rich>=14.3.2",
]
